requirements.txt
streamlit
python-dotenv
Pillow
requests
PyMuPDF
google-generativeai
